Add support for vendoring deps in v7.3.0
This commit allows for the use of vendored deps with v7.3.0 of this client.
It also removes module support and rewrites local imports to properly
reference the yext fork.

Steps followed:
1. Checkout appropriate tag (v7.3.0)
2. go mod vendor in cmd/internal/generate
3. `find . -name "go.mod" -type f -delete`
4. `find . -name "go.sum" -type f -delete`
5. `find . -type f -name "*.go" -exec sed -i '' -e 's#github.com/elastic/go-elasticsearch/v7#github.com/yext/go-elasticsearch#g' {} \;`

Tested this change by running a local test program in which a client
connected to a locally running elasticserver and printed out
client/server versions.